(GH-22) Move Document Store to own module
Previously the Document Store module was located in the message_router.  This
commit moves this into it's own module.
